Move DecodeParams to its own header file.
