+++ include/exiv2/exif.hpp
@@ -16,6 +16,7 @@
 // included header files
 #include "config.h"
 #include "metadatum.hpp"
+#include "params.hpp"
 
 // + standard includes
 #include <list>
@@ -477,26 +478,6 @@ class EXIV2API ExifData {
   // DATA
   ExifMetadata exifMetadata_;
 };  // class ExifData
-
-/*!
-  @brief Parameters for the "decode" functions. There are a fairly large
-  number of static "decode" functions. Examples are `ExifParser::decode`,
-  `TiffParser::decode`, and `XmpParser::decode`. Similar to `ImageCtorParams` above,
-  this class is a common set of parameters for those functions. It currently
-  only contains a `max_recursion_depth_` field, but it will make it easier to
-  add new parameters in the future.
- */
-class EXIV2API DecodeParams {
- public:
-  explicit DecodeParams(size_t max_recursion_depth);
-
-  size_t max_recursion_depth() const {
-    return max_recursion_depth_;
-  }
-
- private:
-  const size_t max_recursion_depth_;
-};
 
 /*!
   @brief Stateless parser class for Exif data. Images use this class to
